Full move legality checking
The analyze_move analyzes any particular move
for how many tiles would be flipped in the 8
possible directions by using analyze_direction.

The mark_possible_moves uses the analyze_move
fn to find all moves that result in a score
that is greater than 0, and then marks them
on the board.

These functions are untested. The move
marking is especially helpful for beginners.